minor - checks before deleting users in Admin users list

+++ frontend/src/screens/UsersScreen.jsx
@@ -5,25 +5,34 @@ import {toast} from 'react-toastify'
 import IdleTimeout from "../components/IdleTimeout"
 import useIdleLogout from "../hooks/useIdleLogout"
 import {useFetchUsersQuery, useDeleteUserMutation} from "../store/slices/usersApiSlice"
+import { useFetchAccountsQuery } from '../store/slices/accountApiSlice';
 import Col from 'react-bootstrap/Col';
 import Row from 'react-bootstrap/Row';
 import AppModal from '../components/AppModal';
 import UserEdit from '../components/UserEdit';
 import UserAdd from '../components/UserAdd';
-import useGetAccountFromId from '../hooks/useGetAccountFromId';
+import { useSelector } from 'react-redux'
+
 
 export default function UsersScreen(){
 
 
 
 
-    const {data, error, isLoading} = useFetchUsersQuery()
+    const {data:users, error, isLoading} = useFetchUsersQuery()
+    const {data:accounts, error:errorAccounts, isLoading:isLoadingAccounts} = useFetchAccountsQuery()
     const [deleteUser, results]=useDeleteUserMutation()
+    const {userInfo:loggedInUser} = useSelector(state=>state.auth)
 
 
     
 
     const handleX=async (userId)=>{
+
+
+            if(userId===loggedInUser._id){
+                toast.error('You cannot delete your own user!') 
+                return}
             const res= await deleteUser(userId)
             !res.error?toast.success("User deleted successfully"):toast.error(res.error?.data?.message)
             // console.log(res)
@@ -39,8 +48,10 @@ export default function UsersScreen(){
     } else {
         content = (
             <div>
-                {data.map(user => {
-
+                {users.map(user => {
+                    if(errorAccounts||!accounts) {return}
+                    
+                    const account= accounts.filter(account=>account._id===user.account)
                     return(
 
 
@@ -50,7 +61,7 @@ export default function UsersScreen(){
                     <Col sm={2}>{user.name}</Col>
                     <Col sm={2}>{user.email}</Col>
                     <Col sm={2}>{user.role||null}</Col>
-                    <Col sm={3}>{user.account||null}</Col>
+                    <Col sm={3}>{account[0].number||null}</Col>
                     {/* <Col sm={3}>{accountId||null}</Col> */}
                    
                     <Col sm={2}>
